% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_state-trace.R
\name{cmr}
\alias{cmr}
\title{Coupled Monotonic Regression}
\usage{
cmr(
  data,
  col_value,
  col_participant,
  col_dv,
  col_within,
  col_between,
  partial,
  test = TRUE,
  nsample = 1000,
  shrink = -1,
  approx = FALSE,
  tolerance = 1e-04
)
}
\arguments{
\item{data}{\code{data.frame} containing data aggregated by participant and
relevant variables in columns.}

\item{col_value}{\code{character}. Name of column in \code{data} containing numerical
values for analysis (i.e., responses).}

\item{col_participant}{\code{character}. Name of column in \code{data} containing the
participant identifier.}

\item{col_dv}{\code{character}. Name of column in \code{data} containing the dependent
variable(s) spanning the state-trace axes.}

\item{col_within}{\code{character}, optional. Name of column(s) in \code{data}
containing the within-subjects variables.}

\item{col_between}{\code{character}, optional. Name of column(s) in \code{data}
containing the between-subjects variables.}

\item{partial}{defines a partial order. Either a \code{character} or a named
\code{list} of characters. See details for ways to specify a partial order.}

\item{test}{logical. If \code{TRUE} (the default) \emph{p}-value is calculated based
double-bootsrap procedure with \code{nsamples}. If \code{FALSE}, no test statistic is
approximated and model is only fit.}

\item{nsample}{number of bootstrap samples to empirically approximate the
null distribution (default is 1000, but about 10000 is probably better).
Only used if \code{test = TRUE}.}

\item{shrink}{Shrinkage parameter (see \link{sta_stats}). Default calculates
optimum amount of shrinkage.}

\item{approx}{\code{FALSE} (the default) uses full algorithm, \code{TRUE} uses an
approximate algorithm that should be used for large problems.}

\item{tolerance}{tolerance used during optimization for numerical stability
(function values smaller than \code{tolerance} are set to 0)}
}
\description{
\code{cmr} is the main function that conducts the CMR (state-trace) analysis for
continuous data. It takes the data as a \code{data.frame} and an optional partial
order and returns a fitted model object of class \code{stacmr}. It fits the
conjoint monotonic model to the data and calculates the \emph{p}-value.
}
\examples{

data(delay)
# Fit and test CMR State-Trace Analysis Model
st_d1 <- cmr(
  data = delay,
  col_value = "pc",
  col_participant = "participant",
  col_dv = "structure",
  col_within = "block",
  col_between = "delay",
  nsample = 1e3
)
st_d1  # basic information about coupled-monotonic model

summary(st_d1)  # basic information plus estimated cell means
str(st_d1, 1, give.attr = FALSE) # overview of information in fitted object

# same model with approximate method gives same result here
st_d2 <- cmr(
  data = delay,
  col_value = "pc",
  col_participant = "participant",
  col_dv = "structure",
  col_within = "block",
  col_between = "delay",
  approx = TRUE,
  nsample = 1e3
)
summary(st_d2)

# for delay data: order of factor-levels corresponds to expected partial order.
# Therefore, 'partial = "auto"' can be used to enforce this order.
st_d3 <- cmr(
  data = delay,
  col_value = "pc",
  col_participant = "participant",
  col_dv = "structure",
  col_within = "block",
  col_between = "delay",
  partial = "auto"
)
st_d3
summary(st_d3)

}
