% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacmr-package.R
\docType{package}
\name{stacmr-package}
\alias{stacmr}
\alias{stacmr-package}
\title{stacmr: Coupled Monotonic Regression for State-Trace Analysis}
\description{
State-trace analysis (STA; Bamber, 1979, \doi{10.1016/0022-2496(79)90016-6}) is a general method to determine properties of the latent structure that mediates the effects of one or more independent variables on two or more dependent variables. 'stacmr' implements the Coupled Monotonic Regression approach to State-Trace Analysis described in Dunn and Kalish (2018, \doi{10.1007/978-3-319-73129-2}).
}
\author{
\strong{Maintainer}: Johannes Titz \email{johannes.titz@gmail.com} (\href{https://orcid.org/0000-0002-1102-5719}{ORCID})

Authors:
\itemize{
  \item John Dunn
  \item Luke Finlay
  \item Henrik Singmann \email{singmann@gmail.com} (\href{https://orcid.org/0000-0002-4842-3657}{ORCID})
  \item Michael L. Kalish
}

}
\keyword{internal}
