% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_state-trace.R
\name{cmr}
\alias{cmr}
\alias{continuous_cmr}
\title{CMR State-Trace Analysis}
\usage{
cmr(
  data,
  col_value,
  col_participant,
  col_dv,
  col_within,
  col_between,
  partial,
  test = TRUE,
  nsample = 1000,
  shrink = -1,
  approx = FALSE,
  tolerance = 1e-04
)
}
\arguments{
\item{data}{\code{data.frame} containing data aggregated by participant and
relevant variables in columns.}

\item{col_value}{\code{character}. Name of column in \code{data} containing numerical
values for analysis (i.e., responses).}

\item{col_participant}{\code{character}. Name of column in \code{data} containing the
participant identifier.}

\item{col_dv}{\code{character}. Name of column in \code{data} containing the dependent
variable(s) spanning the state-trace axes.}

\item{col_within}{\code{character}, optional. Name of column(s) in \code{data}
containing the within-subjects variables.}

\item{col_between}{\code{character}, optional. Name of column(s) in \code{data}
containing the between-subjects variables.}

\item{partial}{defines a partial order. Either a \code{character} or a named
\code{list} of characters. See details for ways to specify a partial order.}

\item{test}{logical. If \code{TRUE} (the default) \emph{p}-value is calculated based
double-bootsrap procedure with \code{nsamples}. If \code{FALSE}, no test statistic is
approximated and model is only fit.}

\item{nsample}{number of bootstrap samples to empirically approximate the
null distribution (default is 1000, but about 10000 is probably better).
Only used if \code{test = TRUE}.}

\item{shrink}{Shrinkage parameter (see \link{sta_stats}). Default calculates
optimum amount of shrinkage.}

\item{approx}{\code{FALSE} (the default) uses full algorithm, \code{TRUE} uses an
approximate algorithm that should be used for large problems.}

\item{tolerance}{tolerance used during optimization for numerical stability
(function values smaller than \code{tolerance} are set to 0)}
}
\description{
\code{cmr} is the main function that conducts the CMR (state-trace) analysis for
continuous data. It takes the data as a \code{data.frame} and an optional partial
order and returns a fitted model object of class \code{stacmr}. It fits the
conjoint monotonic model to the data and calculates the \emph{p}-value.

\code{mr} conducts monotonic regression on a data structure according to a
given partial order.
}
\examples{

library("stacmr")

## load data from Exp. 1 of Dunn, Newell, & Kalish (2012)
data(delay)
delay$delay <- factor(delay$delay, levels = c("delay", "no delay"))
#delay$delay <- factor(delay$delay, levels = c("no delay", "delay"))
str(delay, width = 78, strict.width = "cut")
# 'data.frame':	520 obs. of  5 variables:
#  $ participant: Factor w/ 130 levels "1","2","3","4",..: 1 2 3 4 5 6 7 8 9 1..
#  $ delay      : Factor w/ 2 levels "no delay","delay": 2 1 1 2 2 1 1 2 2 2 ...
#  $ structure  : Factor w/ 2 levels "rule-based","information-integration": 1..
#  $ block      : Factor w/ 4 levels "B1","B2","B3",..: 1 1 1 1 1 1 1 1 1 1 ...
#  $ pc         : num  0.338 0.287 0.525 0.35 0.237 ...

stats <- sta_stats(
  data = delay, 
  col_value = "pc", 
  col_participant = "participant",
  col_dv = "structure", 
  col_within = "block", 
  col_between = "delay"
)
stats
str(stats)
summary(stats)

### Fit and test CMR State-Trace Analysis Model
st_d1 <- cmr(
  data = delay, 
  col_value = "pc", 
  col_participant = "participant",
  col_dv = "structure", 
  col_within = "block", 
  col_between = "delay", 
  nsample = 1e4
)
st_d1  ## basic information about conjoint-monotonic model

summary(st_d1)  ## full summary information about conjoint-monotonic model

state_trace(st_d1) ## produces state trace plot

plot_null(st_d1)  ## produces histogram of empirical (boot-strapped) null distribution

str(st_d1, 1, give.attr = FALSE) ## overview of information in fitted object

st_d2 <- cmr(
  data = delay, 
  col_value = "pc", 
  col_participant = "participant",
  col_dv = "structure", 
  col_within = "block", 
  col_between = "delay", 
  approx = TRUE, 
  nsample = 1e4
)

st_d3 <- cmr(
  data = delay, 
  col_value = "pc", 
  col_participant = "participant",
  col_dv = "structure", 
  col_within = "block", 
  col_between = "delay", test = FALSE,
  approx = TRUE, 
  nsample = 1e4
)

### with partial order

## wrong model:
st_d4 <- cmr(
  data = delay, 
  col_value = "pc", 
  col_participant = "participant",
  col_dv = "structure", 
  col_within = "block", 
  col_between = "delay", 
  nsample = 1e4, 
  partial = "auto"
)
st_d4

mr_d4 <- mr(
  data = delay, 
  col_value = "pc", 
  col_participant = "participant",
  col_dv = "structure", 
  col_within = "block", 
  col_between = "delay", 
  nsample = 1e4, 
  partial = "auto"
)
mr_d4

mr_d5 <- mr(
  data = delay, 
  col_value = "pc", 
  col_participant = "participant",
  col_dv = "structure", 
  col_within = "block", 
  col_between = "delay", 
  nsample = 1e4, 
  partial = list(
    delay = "delay < `no delay`",
    block = "B1 < B2 < B3 < B4"
  )
)
mr_d5

mr_d6 <- mr(
  data = delay, 
  col_value = "pc", 
  col_participant = "participant",
  col_dv = "structure", 
  col_within = "block", 
  col_between = "delay", 
  nsample = 1e4, 
  partial = list(
    delay = "delay = `no delay`",
    block = "B1 < B2 < B3 < B4"
  )
)
summary(mr_d6)

mr_d7 <- mr(
  data = delay, 
  col_value = "pc", 
  col_participant = "participant",
  col_dv = "structure", 
  col_within = "block", 
  col_between = "delay", 
  nsample = 1e4, 
  partial = list(
    delay = "`no delay` < delay",
    block = "B1 < B2 < B3 < B4"
  )
)
summary(mr_d7)

mr_d6 <- mr(
  data = delay, 
  col_value = "pc", 
  col_participant = "participant",
  col_dv = "structure", 
  col_within = "block", 
  col_between = "delay", 
  nsample = 1e4, 
  partial = list(
    delay = "delay = `no delay`",
    block = "B1 < B2 < B3 < B4"
  )
)
summary(mr_d6)


### OLD STUFF BELOW: CHANGE PLEASE

#### Fit and specify partial order that should be obeyed by data
# Create Partial Order
E <- list(c(1:4),c(5:8),c(5,1),c(6,2),c(7,3),c(8,4))
list2adj(1:8, E)

### Fits Partial Order Model to Data
fit2 <- fit_mr(data=delay, col_value = "pc", 
               col_participant = "participant",
               col_dv = "structure", 
               col_within = "block", 
               col_between = "delay",
               partial=E)
cbind(fit2$x[[1]],fit2$x[[2]]) # simplify presentation
fit2$fval
fit2$shrinkage

### Test fit of Partial Order Model
test2 <- test_mr(data=delay, col_value = "pc", 
                 col_participant = "participant",
                 col_dv = "structure", 
                 col_within = "block", 
                 col_between = "delay", 
                 partial=E, nsample=10000)
test2$p
test2$datafit

#### test conjoint model versus partial order model
### p-Value of Difference in Fit of Conjoint Monotonic and Fit of Partial Order Model
test3 <- test_cmr(data=delay, col_value = "pc", 
                  col_participant = "participant",
                  col_dv = "structure", 
                  col_within = "block", 
                  col_between = "delay", 
                  partial=E, nsample=10000)
test3$p
test3$datafit


# staPLOT(data = delay, 
#         groups = list(c(1:4), c(5:8)), 
#         grouplabels = list("No delay", "Delay"), 
#         axislabels = list("RB","II"),
#         pred=out1$x)






}
